Clear clipboard 20 times fix (#423)
* Fix clear clipboard 20 times

As described in issue #419 there was a bug when clearing clipboard 20 times because from "0" to "< 19" there are 19 numbers, not 20. 19 is never reached because of "less than" operator. This could be also fixed by changing the operator from "=" to "<=" but I think is better to change the number of times to 20.

* Fixed clear clipboard 20 times hint

Fixed the hint to match what really happens, consecutive numbers are stored in the clipboard 20 times.