Fix deletion of individual password files
Commit fde8137b (#659) introduced a regression that results in Password
Store crashing when the user tries to delete a single password file as
opposed to a directory.

The root cause is a call of FileUtils.listFiles() on the selected item,
which only works for directories.

The fix is to work with a list consisting only of the selected item if
it happens to be a file.