UserPreference: fix my inability to count
In 98e9f6734a00c I changed a check for split.size > 0 to be
split.isNotEmpty, which was technically correct for the condition,
but as a whole made no sense at all, because Kotlin is a real programming
language and thus arrays start at 0. Rather than use isNotEmpty(), the
right thing to do earlier should have been this.

Signed-off-by: Harsh Shandilya <me@msfjarvis.dev>