Wrap password creation layout in ScrollView
Previously, some of the inputs in the PasswordCreationActivity would be
off-screen and inaccessible, when the screen height was reduced by
rotating the device from portrait to landscape orientation or opening
the on-screen keyboard, for example. The visible portion of the inputs
could not be changed.

To fix this, make the PasswordCreationActivity's layout scrollable. This
way, inputs that are off-screen can be brought into view by scrolling.

Especially the 'Extra content' input is way more accessible now, as it
would also be off-screen previously, even when it was being focused
for editing. Now, it will be automatically brought into view again after
the on-screen keyboard opens.