Add Activity to view the git commit log
The GitLogActivity is accessible via the action bar menu ("View commit
log") if the password repo is a git repository. The activity displays
basic information about the commits in the repository using a
RecyclerView, after internally performing a git-log command. This
information is 1. the short commit hash, 2. the commit's short message,
3. the commit date and time. Other information is omitted from the UI
as of this commit.