Mark unfocused single password fields as fillable (#1258)
Unfocused single password fields in forms with no other recognized
fields must be marked as fillable or the Autofill service will no
longer be invoked on that form.

This is compatible with the restrictions of single-origin mode as
filling only takes place after the password field has gained focus and
the user has tapped the fill UI.

For an example website where fill UI is not shown without this commit,
open https://amazon.lbb.de and tab the user name field ("Benutzername")
first before focusing the password field.