Add snackbar error message when .gpg-id contains a short hex id
I used the first version of Android-Password-Store several years ago
with a .gpg-id containing a short 8-digit hex id, and this worked fine.
When I installed the latest snapshot, it didn't work: when I went to
decrypt any password, nothing would show up.
It turns out that Android-Password-Store doesn't support short hex ids,
but there is no user feedback when this problem arises.

Add an error message when a .gpg-id file with a short hex id is found.